Create class for handle capture responding for BioMini
